Refactor header setup and add download asset test
Updated `SetupHeaders` in `ApiService.cs` to conditionally set the "Accept" header based on the `download` parameter. Added a call to `SetupHeaders(download: true)` in `ReleaseService.cs` to prepare for asset downloads and log the headers. Introduced a new test method `DownloadPublicAssetByName_ShouldDownloadAsset` in `ReleaseServiceTests.cs` to validate the downloading of public assets.