Refactor Clone method to use Cli options
Updated the Clone method in Command.cs to accept a Cli object, allowing for additional options like verbosity. Adjusted the git_clone command in Program.cs to align with this new method signature. Added a verbose output in GitWrapper.cs to provide feedback on the clone path when verbosity is enabled.